# apis/sentinel/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
...
        "title": "Sentinel API",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "0.1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/v1/permissions/{context_id}/resources": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "List all Permissions for all Target Resources in context.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Permissions"
                ],
                "summary": "List all Permissions for all Target Resources in context.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Desired environment",
                        "name": "x-sentinel-tenant",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Claimant requesting the operation",
                        "name": "claimant",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "context ID",
                        "name": "context_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/permission.Output"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/views.ErrView"
                        }
                    }
                }
            }
        },
        "/v1/permissions/{context_id}/resources/{resource_id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "List all permissions for a context for a given target",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Permissions"
                ],
                "summary": "List all permissions for a context for a given target",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Desired environment",
                        "name": "x-sentinel-tenant",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Claimant requesting the operation",
                        "name": "claimant",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "context ID",
                        "name": "context_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Resource ID",
                        "name": "resource_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/permission.Output"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/views.ErrView"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Create a permission for a context on a resource",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Permissions"
                ],
                "summary": "Update a permission that permits acontext on a resource",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Desired environment",
                        "name": "x-sentinel-tenant",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Claimant requesting the operation",
                        "name": "claimant",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "context ID",
                        "name": "context_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Resource ID",
                        "name": "resource_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Details about the permission to be added",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/permission.Input"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/permission.OutputDetails"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/views.ErrView"
                        }
                    }
                }
            }
        },
        "/v1/permissions/{id}": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Delete a permission by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Permissions"
                ],
                "summary": "Delete a permission by its ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Desired environment",
                        "name": "x-sentinel-tenant",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Claimant requesting the operation",
                        "name": "claimant",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Permission ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/views.ErrView"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Update Permission Details.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Permissions"
                ],
                "summary": "Update a Permission by its ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Desired environment",
                        "name": "x-sentinel-tenant",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Claimant requesting the operation",
                        "name": "claimant",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Permission ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "New name to be assigned to an existing permission",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/permission.Input"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/permission.OutputDetails"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/views.ErrView"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "permission.Attributes": {
            "type": "object",
            "required": [
                "name",
                "permitted"
            ],
            "properties": {
                "name": {
                    "type": "string"
                },
                "permitted": {
                    "type": "string"
                }
            }
        },
        "permission.Details": {
            "type": "object",
            "required": [
                "attributes",
                "type"
            ],
            "properties": {
                "attributes": {
                    "type": "object",
                    "$ref": "#/definitions/permission.Attributes"
                },
                "id": {
                    "type": "string"
                },
                "type": {
                    "type": "string",
                    "enum": [
                        "permission"
                    ]
                }
            }
        },
        "permission.Input": {
            "type": "object",
            "required": [
                "data"
            ],
            "properties": {
                "data": {
                    "type": "object",
                    "$ref": "#/definitions/permission.InputDetails"
                }
            }
        },
        "permission.InputDetails": {
            "type": "object",
            "required": [
                "attributes",
                "type"
            ],
            "properties": {
                "attributes": {
                    "type": "object",
                    "$ref": "#/definitions/permission.Attributes"
                },
                "type": {
                    "type": "string",
                    "enum": [
                        "permission"
                    ]
                }
            }
        },
        "permission.Output": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/permission.Details"
                    }
                }
            }
        },
        "permission.OutputDetails": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object",
                    "$ref": "#/definitions/permission.Details"
                }
            }
        },
...
        "views.ErrView": {
            "type": "object",
            "properties": {
                "detail": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "source": {
                    "type": "object",
                    "$ref": "#/definitions/views.Source"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "views.Source": {
            "type": "object",
            "properties": {
                "parameter": {
                    "type": "string"
                },
                "pointer": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}